+++ frontend/jupyter/src/app/pages/index/index-default/index-default.component.html
@@ -10,5 +10,5 @@
 <lib-resource-table
   [config]="costConfig"
   [data]="processedCostData"
-
+  [trackByFn]="costTrackByFn"
 ></lib-resource-table>
