+++ frontend/jupyter/src/app/services/backend.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { BackendService, SnackBarService } from 'kubeflow';
-import { HttpClient, HttpErrorResponse } from '@angular/common/http';
-import { Observable, throwError } from 'rxjs';
-import { catchError, map, tap } from 'rxjs/operators';
+import { HttpClient } from '@angular/common/http';
+import { Observable } from 'rxjs';
+import { catchError, map } from 'rxjs/operators';
 import {
   NotebookResponseObject,
   JWABackendResponse,
@@ -128,71 +128,5 @@ export class JWABackendService extends BackendService {
     return this.http
       .delete<JWABackendResponse>(url)
       .pipe(catchError(error => this.handleError(error, false)));
-  }
-}
-
-export type AggregateCostResponse = {
-  code: number;
-  data: {
-    [namespace: string]: {
-      aggregation: string;
-      environment: string;
-      cpuAllocationAverage: number;
-      cpuCost: number;
-      cpuEfficiency: number;
-      efficiency: number;
-      gpuAllocationAverage: number;
-      gpuCost: number;
-      ramAllocationAverage: number;
-      ramCost: number;
-      ramEfficiency: number;
-      pvAllocationAverage: number;
-      pvCost: number;
-      networkCost: number;
-      sharedCost: number;
-      totalCost: number;
-    };
-  };
-  message: string;
-};
-
-@Injectable()
-export class KubecostService {
-  constructor(private http: HttpClient) {}
-
-  getAggregateCost(ns: string): Observable<AggregateCostResponse> {
-    const url = `/api/namespaces/${ns}/cost/aggregated`;
-
-    return this.http
-      .get<AggregateCostResponse | JWABackendResponse>(url, {
-        params: {
-          aggregation: "namespace",
-          namespace: ns,
-          window: "1d"
-        }
-      })
-      .pipe(
-        tap(res => this.handleBackendError(res)),
-        catchError(err => this.handleError(err))
-      ) as Observable<AggregateCostResponse>;
-  }
-
-  private handleBackendError(response: AggregateCostResponse | JWABackendResponse) {
-    if (this.isResp(response) || response.code < 200 || response.code >= 300) {
-      throw response;
-    }
-  }
-
-  private handleError(
-    error: HttpErrorResponse | AggregateCostResponse | JWABackendResponse
-  ): Observable<never> {
-    const message = this.isResp(error) ? error.log : error.message;
-    return throwError(new Error(message));
-  }
-
-  private isResp(
-    obj: HttpErrorResponse | AggregateCostResponse | JWABackendResponse
-  ): obj is JWABackendResponse {
-    return (obj as JWABackendResponse).success !== undefined;
   }
 }
