+++ frontend/jupyter/src/app/services/kubecost.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { KubecostService } from './kubecost.service';
+
+describe('KubecostService', () => {
+  let service: KubecostService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(KubecostService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
